refactor: restructure modules and add logging support
