+++ assets_to_usd/src/assets_to_usd/asset_mapping.py
