+++ assets_to_usd/src/assets_to_usd/texture_resolve.py
@@ -1,7 +1,7 @@
 import pathlib
 from dataclasses import dataclass, field
 
-from assets_to_usd._utils import flatten_input_data
+from assets_to_usd.asset_mapping import flatten_input_data
 
 
 @dataclass
@@ -21,8 +21,9 @@ class TextureResolve:
     mapping: dict[str, pathlib.Path] = field(default_factory=dict)
 
     def parse_texture(self) -> dict:
-        # we remove the first few characters just because we use asset name as a namespace, and sometimes it comes with prefixes
-        name_space_edit = flatten_input_data(self.namespace)[5:]
+        # we remove the first few characters because it comes with prefix
+        name_space_clean = flatten_input_data(self.namespace)
+        name_space_edit = name_space_clean[5:] if len(name_space_clean) > 5 else name_space_clean
 
         tex_folder_abs_path = pathlib.Path(self.tex_folder_path)
         for file in tex_folder_abs_path.iterdir():
